Clean up config logic
