+++ data/config/__init__.py
@@ -2,17 +2,17 @@ import json
 import os
 import sys
 
-from datetime import datetime, time, timedelta
-from enum import Enum
+from datetime import datetime, timedelta
 from collections import defaultdict
-from typing import Mapping, Optional
+from typing import Mapping
 from math import ceil
 
+from data.config.game_screen import GameScreen, parse_game_screen
+from data.config.other_screens import VALID_NON_GAME_SCREEN_TYPES, TimeRule, parse_time_rule, parse_with_priority
 import debug
 from data import status
 from data.config.color import Color
 from data.config.layout import Layout
-from data import teams as team_metadata
 from data.paths import *
 from data.time_formats import TIME_FORMAT_12H, TIME_FORMAT_24H, os_datetime_format
 from utils import deep_update
@@ -29,8 +29,6 @@ DEFAULT_PREFERRED_DIVISIONS = ["NL Central"]
 class Config:
     def __init__(self, config_path, width, height):
         json = self.__get_config(config_path)
-
-        # Preferred Teams/Divisions
 
         # News Ticker
         self.preferred_teams = json["news_ticker"]["teams"]
@@ -112,15 +110,10 @@ class Config:
 
     def check_delay(self):
         if self.sync_delay_seconds < 0:
-            debug.warning(
-                "sync_delay_seconds should be a positive integer. Using default value of 0"
-            )
+            debug.warning("sync_delay_seconds should be a positive integer. Using default value of 0")
             self.sync_delay_seconds = 0
         if self.sync_delay_seconds != int(self.sync_delay_seconds):
-            debug.warning(
-                "sync_delay_seconds should be an integer."
-                f" Truncating to {int(self.sync_delay_seconds)}"
-            )
+            debug.warning("sync_delay_seconds should be an integer." f" Truncating to {int(self.sync_delay_seconds)}")
             self.sync_delay_seconds = int(self.sync_delay_seconds)
 
     def check_api_refresh_rate(self):
@@ -300,149 +293,7 @@ If you aren't sure why you're seeing this, there might not be official support f
         return vars(self) == vars(other)
 
 
-class Requirements(Enum):
-    LIVE = "live"
-    LIVE_IN_INNING = "live_in_inning"
-    PREGAME = "pregame"
-    GAME_OVER = "game_over"
-
-    def __str__(self):
-        return self.value
-
-    @staticmethod
-    def from_str(label):
-        for requirement in Requirements:
-            if requirement.value == label:
-                return requirement
-        raise ValueError(f"Unknown requirement: {label}")
-
-
-class GameRule:
-    DEFAULT_PRIORITY = 0, True
-
-    def __init__(
-        self,
-        priority: int,
-        *,
-        requirement: Optional[Requirements] = None,
-        passive=False,
-        teams: list[str] = [],
-    ):
-        self.requirement = requirement
-        self.when_matched = priority, passive
-        self.teams = set(team_metadata.get_team_id(t) for t in teams)
-
-    def priority(self) -> int:
-        return self.when_matched[0]
-
-    def matches(self, game) -> tuple[int, bool]:
-        if self.teams and not set([game["away_id"], game["home_id"]]).intersection(self.teams):
-            return GameRule.DEFAULT_PRIORITY
-
-        if self.requirement is None:
-            return self.when_matched
-
-        if self.requirement == Requirements.PREGAME and status.is_pregame(game["status"]):
-            return self.when_matched
-
-        if self.requirement == Requirements.GAME_OVER and status.is_complete(game["status"]):
-            return self.when_matched
-
-        if self.requirement == Requirements.LIVE and (
-            status.is_fresh(game["status"]) or (status.is_live(game["status"]))
-        ):
-            return self.when_matched
-
-        if self.requirement == Requirements.LIVE_IN_INNING and (
-            status.is_live(game["status"])
-            and game["status"] != status.WARMUP
-            and not status.is_inning_break(game["inning_state"])
-        ):
-            return self.when_matched
-
-        return GameRule.DEFAULT_PRIORITY
-
-    def __repr__(self):
-        return (
-            f"GameRule(priority={self.when_matched[0]}, requirement={self.requirement}"
-            f", passive={self.when_matched[1]}, teams={self.teams})"
-        )
-
-    def __eq__(self, other):
-        if not isinstance(other, GameRule):
-            return NotImplemented
-        return (
-            self.requirement == other.requirement
-            and self.when_matched == other.when_matched
-            and self.teams == other.teams
-        )
-
-
-class TimeRule:
-    # TODO(bmw): extend to day of week?
-    def __init__(
-        self,
-        priority: int,
-        *,
-        start_time: Optional[time] = None,
-        end_time: Optional[time] = None,
-    ):
-        self.priority = priority
-        self.start_time = start_time
-        self.end_time = end_time
-
-    def matches(self, now: time) -> int:
-        if self.start_time and now < self.start_time:
-            return 0
-        if self.end_time and now > self.end_time:
-            return 0
-        return self.priority
-
-    def __repr__(self):
-        return f"TimeRule(priority={self.priority}, start_time={self.start_time}, end_time={self.end_time})"
-
-    def __eq__(self, other):
-        if not isinstance(other, TimeRule):
-            return NotImplemented
-        return (
-            self.priority == other.priority
-            and self.start_time == other.start_time
-            and self.end_time == other.end_time
-        )
-
-def _parse_requirements(json) -> Optional[Requirements]:
-    json_requirement = json.get("required_status")
-    if json_requirement:
-        try:
-            return Requirements.from_str(json_requirement)
-        except ValueError:
-            raise ValueError(
-                "Invalid game rule in config, unknown required_status '{}'. Rule: {}".format(json_requirement, json)
-            )
-    return None
-
-
-def _parse_with_priority(json) -> list[int]:
-    with_priority = json.get("with_priority")
-    if with_priority is None:
-        raise ValueError("Invalid screen rule in config, missing 'with_priority' field. Rule: {}".format(json))
-
-    if isinstance(with_priority, int):
-        return [with_priority]
-    elif isinstance(with_priority, list) and all(isinstance(p, int) for p in with_priority):
-        return with_priority
-    else:
-        raise ValueError(
-            "Invalid screen rule in config, 'with_priority' field should be an integer or list of integers. Rule: {}".format(
-                json
-            )
-        )
-
-
-VALID_NON_GAME_SCREEN_TYPES = ["news", "standings"]
-
-
-def _screen_rules_from_json(json) -> tuple[list[GameRule], list[TimeRule], Mapping[str, Mapping[int, int]]]:
+def _screen_rules_from_json(json) -> tuple[list[GameScreen], list[TimeRule], Mapping[str, Mapping[int, int]]]:
     game_rules = []
     time_rules = []
     screen_rules: defaultdict[str, defaultdict[int, int]] = defaultdict(lambda: defaultdict(int))
@@ -451,57 +302,14 @@ def _screen_rules_from_json(json) -> tuple[list[GameRule], list[TimeRule], Mappi
         if "kind" not in rule_json:
             raise ValueError("Invalid rule in config, missing 'kind' field. Rule: {}".format(rule_json))
 
-        if rule_json["kind"] == "game":
-            if "priority" not in rule_json:
-                raise ValueError("Invalid game rule in config, missing 'priority' field. Rule: {}".format(rule_json))
-            rule = GameRule(
-                priority=rule_json["priority"],
-                requirement=_parse_requirements(rule_json),
-                passive=False,
-                teams=rule_json.get("teams", []),
-            )
-            game_rules.append(rule)
-        elif rule_json["kind"] == "secondary_game":
-            requirement = _parse_requirements(rule_json)
-            for priority in _parse_with_priority(rule_json):
-                rule = GameRule(
-                    priority=priority,
-                    requirement=requirement,
-                    passive=True,
-                    teams=rule_json.get("teams", []),
-                )
-                game_rules.append(rule)
+        if rule_json["kind"] == "game" or rule_json["kind"] == "secondary_game":
+            game_rules.extend(parse_game_screen(rule_json))
         elif rule_json["kind"] == "time":
-            if "priority" not in rule_json:
-                raise ValueError("Invalid time rule in config, missing 'priority' field. Rule: {}".format(rule_json))
-            start_time = None
-            end_time = None
-            if "start_time" in rule_json:
-                try:
-                    start_time = datetime.strptime(rule_json["start_time"], "%H:%M").time()
-                except ValueError:
-                    raise ValueError(
-                        "Invalid time format for 'start_time' in config. Expected HH:MM. Rule: {}".format(rule_json)
-                    )
-            if "end_time" in rule_json:
-                try:
-                    end_time = datetime.strptime(rule_json["end_time"], "%H:%M").time()
-                except ValueError:
-                    raise ValueError(
-                        "Invalid time format for 'end_time' in config. Expected HH:MM. Rule: {}".format(rule_json)
-                    )
-            if start_time is None and end_time is None:
-                raise ValueError(
-                    "Invalid time rule in config, need at least one of 'start_time' or 'end_time' fields. Rule: {}".format(
-                        rule_json
-                    )
-                )
-            time_rules.append(TimeRule(priority=rule_json["priority"], start_time=start_time, end_time=end_time))
-
+            time_rules.append(parse_time_rule(rule_json))
         elif rule_json["kind"] in VALID_NON_GAME_SCREEN_TYPES:
             if "seconds" not in rule_json:
                 raise ValueError("Invalid screen rule in config, missing 'seconds' field. Rule: {}".format(rule_json))
-            for priority in _parse_with_priority(rule_json):
+            for priority in parse_with_priority(rule_json):
                 screen_rules[rule_json["kind"]][priority] = rule_json["seconds"]
         else:
             debug.warning(
